Move dontTouch out of experimental package.
